refactor config center
